Modify TernaryDataTransform to be callable from Axes
The issue was that the previous TernaryDataTransfrom could not follow
the change of the ax.bbox. The temtative solution is that the relevant
ternary-related transforms have e.g. transAxes or transLimits
"internally" and when necessary the ternary-related transforms call
them inside the "transform" method. Then the updated information can be
naturally used.