Modify the algorithm for the axis-label rotation and alignment
In the previous commit (be39512c2c473be491e7d045ce7fd1b0ce7505c5), the
algorithm used the display coordinates of the axis label. This, however,
made the testing difficult, because the label coordinates were
determined only after the figure drawing. Now the algorithm is modified
so that it uses only the corner positions of the triangle.