Fix the algorithm for the axis-label rotation and alignment
The previous commit (c22e0e788095afe9934ba0aa43f032f9d68db0b2) did not
work properly for non-regular triangles. Now the information of
label_rotation, axis_angle, clockwiseness of the triangle, and whether
the axis-label is on the corner or not is used to determine the
axis-label rotation.