Refactor TernaryPerpendicularTransform
In the previous `VerticalTernaryTransform`, the y-coordinate direction
was simply counterclockwise to the axis direction. This, however, made
the axis-label positioning complecated due to the case devision whether
the triagnle is clockwise of counterclockwise. In the new
`TernaryPerpendicularTransform`, the y-coordinate direction is always
inside the triangle. This makes the axis-label positioning much simpler.