Add `transTernaryAxes` and `TransAxesProjection`
By setting `ax.transTernaryAxes` as the `transform` argument, we can
plot in the fixed barycentric coordinates, which may be useful
particularly for interactive modes.
Due to the current limit of Matplotlib, we cannot pass
`ax.transTernaryAxes` directly to the original Matplotlib plotting
methods. Therefore, when `ax.transTernaryAxes` is passed, internally
`mpltern` replaces it with `ax.TransAxes` and provide the Axes
coordinates obtained by `TransAxesProjection`.
This implementation is currently done only for `ax.text` and `ax.arrow`.
To aplly this treatment systematic for other methods, some general
methods (maybe as decorators?) should be created later.