Fix to update transforms of tick labels when locations are updated
Previously the pads of tick labels from the axis and the tick markers
were determined when the `Text` objects for tick labels are created and
never updated afterwords. Actually when drawing the triangle geometry is
updated and hence the pads should be also modified following the change
of the triangle geometry. Now this is fixed, and the pads become
consistent for all the axes.