Modify the limits of the original data coordinates
Previously both the data and the `Axes` coordinates had the same range,
[0, 1]. While this was useful to simplify the algorithms, this put a
legend on a position a bit close to the triangle, while might look
inconsistent with normal Matplotlib plots.
Now the x and the y lengths are modified to `1.0 / np.sqrt(3.0)` and
`1.0`, respectively. This should makes the default legend positions just
above the right or the left corner.